backend/requirements.txt
# # BASIX IP-Marketplace: Backend Dependencies

# # Build tools
setuptools<70.0.0  # Pinned to avoid distutils issues with Python 3.12
wheel>=0.43.0  # Ensure latest wheel support

# # Core Flask and web framework
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-Limiter==3.5.0
Flask-Migrate==4.0.5
Werkzeug==2.3.7
Flask-Session==0.5.0
Flask-Caching==2.1.0
flask-restx==1.3.0
Flask-Talisman==1.1.0
Flask-GraphQL==2.0.1
Flask-SocketIO==5.3.6
flask-healthcheck==0.1.2
flask-api-versioning==1.0.0

# # Database and ORM
SQLAlchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9  # PostgreSQL adapter
redis==5.0.1
psycopg2-pool==1.1
sqlalchemy-pool==0.1.0

# # Blockchain and Web3
web3==6.11.3
eth-account==0.9.0
eth-utils==2.2.0
eth-typing==3.5.0
eth-hash==0.5.2
eth-keyfile==0.6.1

# # AI and Machine Learning
numpy>=1.26.0  # Updated for Python 3.12 compatibility
pandas==2.0.3
scikit-learn==1.3.0
tensorflow>=2.16.0  # Updated for Python 3.12 support
torch>=2.3.0  # Updated for Python 3.12 support
scipy==1.11.3
joblib==1.3.2

# # MeTTa Integration
hyperon==0.2.6  # Updated to match installed version
metta-python==0.2.6  # Ensure compatibility with hyperon
metta-core==0.2.6  # Ensure compatibility
metta-hyperon==0.2.6  # Ensure compatibility

# Data validation and serialization
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
pydantic==2.4.2
apispec==6.3.0
marshmallow-apispec==0.5.17

# # Background tasks and job queue
celery==5.3.4
kombu==5.3.4
celery-beat==2.5.0

# HTTP requests and API integration
requests==2.31.0
httpx==0.25.0
aiohttp==3.8.6

# Security and cryptography
cryptography==41.0.7
bcrypt==4.0.1
PyJWT==2.8.0

# File handling and storage
Pillow==10.0.1  # Image processing
python-magic==0.4.27  # File type detection
boto3==1.34.0  # AWS S3 integration
google-cloud-storage==2.10.0  # Google Cloud Storage

# Monitoring and logging
structlog==23.2.0
prometheus-client==0.17.1
sentry-sdk[flask]==1.38.0
rollbar==0.16.3
newrelic==9.1.0

# Testing
pytest==7.4.2
pytest-flask==1.3.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==20.1.0
bandit==1.7.5
safety==2.3.5

# Development and debugging
python-dotenv==1.0.0
black==23.9.1  # Code formatting
flake8==6.1.0  # Linting
mypy==1.6.1  # Type checking
pre-commit==3.5.0
ipython==8.16.1
jupyter==1.0.0
notebook==7.0.2
watchdog==3.0.0

# Documentation
Sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Performance and optimization
gunicorn==21.2.0
uvicorn==0.23.2
asyncio==3.4.3

# Data processing and analysis
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0
dash==2.14.2

# Email and notifications
Flask-Mail==0.9.1  # Removed due to obsolescence; consider Flask-Email or sendgrid
sendgrid==6.10.0

# API documentation
# Already included in flask-restx and apispec

# # Environment and configuration
python-decouple==3.8
click==8.1.7
dynaconf==3.2.4

# # Date and time handling
python-dateutil==2.8.2
pytz==2023.3

# # JSON and data serialization
    orjson==3.9.10
    ujson==5.8.0

# # Network and async
aiofiles==23.2.1
asyncio-mqtt==0.13.0
python-socketio==5.9.0

# # Search functionality
elasticsearch==8.10.0
elasticsearch-dsl==8.11.0

# # Task scheduling
APScheduler==3.10.4

# # Web scraping
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2

# # Performance profiling
memory-profiler==0.61.0
line-profiler==4.1.1

# Internationalization
Flask-Babel==4.0.0
Babel==2.13.1

# Production deployment
supervisor==4.2.5
fabric==2.7.1